almost finished with security

+++ src/main/java/tk/ta4anka/employeemanager/service/UserService.java
@@ -2,6 +2,6 @@ package tk.ta4anka.employeemanager.service;
 
 import tk.ta4anka.employeemanager.model.User;
 
-public interface UserService extends GenericService<User,Long> {
+public interface UserService extends GenericService<User,Integer> {
     User findByUsername(String username);
 }
